+++ packages/component-library/src/components/atoms/Input/index.tsx
@@ -1,12 +1,13 @@
-import type { FC, InputHTMLAttributes } from 'react';
+import type { FC, InputHTMLAttributes, RefObject } from 'react';
 import cs from 'classnames';
 import styles from './Input.module.scss';
 
 export interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
   invalid?: boolean;
+  innerRef?: RefObject<HTMLInputElement>;
 }
 
-export const Input: FC<InputProps> = ({ invalid, className, required, type = 'text', ...rest }) => (
+export const Input: FC<InputProps> = ({ invalid, className, required, innerRef, type = 'text', ...rest }) => (
   <input
     autoComplete="off"
     required={required}
@@ -15,6 +16,7 @@ export const Input: FC<InputProps> = ({ invalid, className, required, type = 'te
       [styles.invalid]: invalid,
     })}
     type={type}
+    ref={innerRef}
     {...rest}
   />
 );