+++ packages/devtools/src/scripts/createBabelConfig.ts
@@ -80,6 +80,7 @@ export const createBabelConfig = ({
         },
       ],
       ['@babel/plugin-proposal-decorators', { legacy: true }],
+      ['@babel/plugin-proposal-private-methods', { loose: true }],
       '@babel/plugin-proposal-numeric-separator',
       [
         '@babel/plugin-transform-runtime',