+++ packages/devtools/tools/scripts/createBabelConfig.js
@@ -71,6 +71,7 @@ var createBabelConfig = function (_a) {
                 },
             ],
             ['@babel/plugin-proposal-decorators', { legacy: true }],
+            ['@babel/plugin-proposal-private-methods', { loose: true }],
             '@babel/plugin-proposal-numeric-separator',
             [
                 '@babel/plugin-transform-runtime',