+++ packages/website/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @cutting/website
 
+## 4.24.11
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @cutting/hooks@4.26.0
+  - @cutting/component-library@4.24.3
+
+## 4.24.10
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @cutting/component-library@4.24.2
+
+## 4.24.9
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @cutting/component-library@4.24.1
+
+## 4.24.8
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @cutting/component-library@4.24.0
+
 ## 4.24.7
 
 ### Patch Changes